chore: update search input style on mobile.
