+++ src/components/navbar.js
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
 
 import { Button } from '@seagreenio/react-bulma'
-import { FormattedMessage, injectIntl } from 'react-intl'
+import { FormattedMessage } from 'react-intl'
 import IntlLink from '../components/IntlLink'
+import SearchInput from './searchInput'
 
-const Navbar = ({ intl }) => {
+const Navbar = () => {
   const { BrandSVG } = useStaticQuery(
     graphql`
       query {
@@ -19,49 +20,6 @@ const Navbar = ({ intl }) => {
   const [showBorder, setShowBorder] = useState(false)
   const [burgerActive, setBurgerActive] = useState(false)
   const handleSetBurgerActive = () => setBurgerActive(!burgerActive)
-  const placeholder = intl.formatMessage({ id: 'navbar.searchDocs' })
-  const versionRegx = /v\d.\d|dev/
-  const [version, setVersion] = useState(null)
-  const [docsType, setDocsType] = useState(null)
-  const [lang, setLang] = useState(null)
-  const [searchQuery, setSearchQuery] = useState(null)
-
-  function getParams() {
-    const searchQuery = new URLSearchParams(window.location.search)
-    let _version = searchQuery.get('version') || ''
-    let _docsType = searchQuery.get('type') || ''
-    const params = window.location.pathname
-    const paramsArr = params.split('/')
-    let _lang = paramsArr[1] === 'zh' ? 'zh/' : ''
-
-    if (!_version || !_docsType) {
-      _version = params.match(versionRegx) ? params.match(versionRegx)[0] : ''
-      switch (_lang) {
-        case 'zh/':
-          _docsType = paramsArr[2]
-          break
-        default:
-          _docsType = paramsArr[1]
-          break
-      }
-    }
-
-    if (_docsType === 'tidb-in-kubernetes') {
-      _docsType = 'tidb-operator'
-    }
-
-    setVersion(_version)
-    setDocsType(_docsType)
-    setLang(_lang)
-  }
-
-  function searchQueryChanged(e) {
-    setSearchQuery(e.target.value)
-  }
-
-  useEffect(() => {
-    getParams()
-  }, [searchQuery])
 
   useEffect(() => {
     const scrollListener = () => {
@@ -92,6 +50,9 @@ const Navbar = ({ intl }) => {
             />
           </IntlLink>
 
+          <div className="navbar-item search-input-mobile">
+            <SearchInput />
+          </div>
           <button
             className={`navbar-burger${burgerActive ? ' is-active' : ''}`}
             aria-label="menu"
@@ -138,23 +99,14 @@ const Navbar = ({ intl }) => {
                 <FormattedMessage id="navbar.contactUs" />
               </Button>
             </div>
-            <form
-              className="navbar-item with-search-input"
-              method="post"
-              action={`/${lang}search?type=${docsType}&version=${version}&q=${searchQuery}`}
-            >
-              <input
-                className="search-input"
-                type="search"
-                placeholder={placeholder}
-                onChange={searchQueryChanged}
-              />
-            </form>
           </div>
+        </div>
+        <div className="navbar-item search-input-pc">
+          <SearchInput />
         </div>
       </div>
     </nav>
   )
 }
 
-export default injectIntl(Navbar)
+export default Navbar
