+++ src/intl/en.json
@@ -19,6 +19,7 @@
     "type": "Docs Type:",
     "version": "Version:",
     "resultTips": {
+      "counts": "{counts} results found...", 
       "title": "Sorry. We couldn't find what you're looking for.",
       "content1": "If you've come to pages of an unexpected language, go to Chinese documentationand try again.",
       "content2": "If you do want to get some English content, PingCAP home page might be a better place for you to go."
